+++ src/standalone.rs
@@ -3,16 +3,11 @@ use crate::args::{
     validate_single_version, versions_argument,
 };
 use crate::container::{
-    container_network, container_run, ensure_unique_names, verify_container_command,
+    container_network, container_run, ensure_unique_names, run_instances, verify_container_command,
 };
-use crate::progress::{Progress, stderr_reader, summary};
 use crate::wildfly::{AdminContainer, Ports, ServerType, StandaloneInstance};
 use clap::ArgMatches;
 use futures::executor::block_on;
-use indicatif::MultiProgress;
-use std::process::Stdio;
-use tokio::task::JoinSet;
-use tokio::time::Instant;
 
 // ------------------------------------------------------ start
 
@@ -55,46 +50,19 @@ async fn start_instances(
     parameters: Vec<String>,
     operations: Vec<String>,
 ) -> anyhow::Result<()> {
-    let count = instances.len();
-    let instant = Instant::now();
-    let multi_progress = MultiProgress::new();
-    let mut commands = JoinSet::new();
-
     container_network().await?;
-    for instance in instances {
-        let progress = Progress::new(
-            &instance.admin_container.wildfly_container.short_version,
-            &instance.admin_container.image_name(),
-        );
-        multi_progress.add(progress.bar.clone());
+    run_instances(&instances, |instance| {
         let mut command = container_run(
-            instance.name.as_str(),
+            &instance.name,
             Some(&instance.ports),
             operations.clone(),
         );
         command
             .arg(instance.admin_container.image_name())
             .args(parameters.clone());
-        let mut child = command
-            .stdout(Stdio::piped())
-            .stderr(Stdio::piped())
-            .spawn()
-            .expect("Unable to run podman-run.");
-
-        let stderr = stderr_reader(&mut child);
-        let progress_clone = progress.clone();
-        commands.spawn(async move {
-            let output = child.wait_with_output().await;
-            progress.finish(output, Some(&instance.name))
-        });
-        tokio::spawn(async move {
-            progress_clone.trace_progress(stderr).await;
-        });
-    }
-
-    let status = commands.join_all().await;
-    summary("Started", "container", count, instant, status);
-    Ok(())
+        command
+    })
+    .await
 }
 
 // ------------------------------------------------------ stop